Updated: Ui till mid-sem review
