+++ app/src/main/res/drawable-v24/background_splash.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <gradient
+        android:angle="90"
+        android:endColor="#734b6d"
+        android:startColor="#42275a" />
+
+</shape>
