+++ app/src/main/res/drawable/background_auth.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <gradient
+        android:angle="-90"
+        android:endColor="#ffde"
+        android:startColor="#ee9ca7" />
+
+</shape>
