--- app/src/main/res/layout/activity_main.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerVertical="true"
-        android:layout_centerHorizontal="true"
-        android:orientation="vertical">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Splash Screen"
-            android:textAlignment="center"
-            android:textSize="20sp"
-            android:textAllCaps="true"/>
-
-        <Button
-            android:id="@+id/btn_splash_skip"
-            android:layout_marginTop="20dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="skip"
-            android:textStyle="bold"
-            android:textAllCaps="true"
-            android:textSize="20sp"
-            android:textAlignment="center"/>
-
-    </LinearLayout>
-
-
-
-</RelativeLayout>
