+++ app/src/main/res/layout/activity_orphanage_dashboard.xml
@@ -1,9 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".orphanage.activities.OrphanageDashboardActivity">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:fontFamily="@font/samarkan"
+        android:text="Work under proccess..!!!!!!"
+        android:textAlignment="center"
+        android:textSize="40sp" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
